move *invites into src/invites/ folders
